app/requirements.txt
flask
gunicorn
redis
psycopg2-binary
requests
